requirements.txt
pytube~=15.0.0
moviepy~=1.0.3
python-slugify~=6.1.2
